chore: shifted blocks provider from shared to builder
