+++ packages/builder/base/src/types/block.ts
@@ -1,0 +1,10 @@
+import type { IconType } from "react-icons/lib";
+
+export type Block<T extends Record<string, unknown> = Record<string, unknown>> =
+  {
+    type: string;
+    label: string;
+    icon: IconType;
+    private?: boolean;
+    presets?: T;
+  };
