+++ packages/builder/canvas/src/form/FormDisplay/FieldWrapper/QuickActions/EditIdField.tsx
@@ -1,7 +1,6 @@
-import { useCanvas } from "@fibr/builder";
+import { useCanvas, classNames, useBoolean, eventHandler } from "@fibr/builder";
 import { useThread } from "@fibr/react";
 import { InputField } from "@rafty/ui";
-import { classNames, useBoolean, eventHandler } from "@fibr/builder";
 import { useEffect, useRef, type FocusEvent, type KeyboardEvent } from "react";
 import { HiPencil } from "react-icons/hi";
 import type { Node } from "reactflow";
