+++ packages/builder/canvas/src/form/FormDisplay/FieldWrapper/index.tsx
@@ -1,8 +1,12 @@
 import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
-import { type BlockType, useCanvas } from "@fibr/builder";
+import {
+  type BlockType,
+  useCanvas,
+  classNames,
+  eventHandler,
+} from "@fibr/builder";
 import { useThread } from "@fibr/react";
-import { classNames, eventHandler } from "@fibr/builder";
 import { cva } from "class-variance-authority";
 import {
   type CSSProperties,
