+++ packages/builder/canvas/src/workflow/NodeWrapper.tsx
@@ -1,6 +1,5 @@
 import { useThread } from "@fibr/react";
-import { useBlocks } from "@fibr/shared";
-import { classNames } from "@fibr/builder";
+import { useBlocks, classNames } from "@fibr/builder";
 import type { PropsWithChildren } from "react";
 import type { Node } from "reactflow";
