+++ packages/builder/shared/src/components/Palette/PaletteCard.tsx
@@ -1,7 +1,11 @@
 import { useDraggable } from "@dnd-kit/core";
-import { type BlockType, classNames, eventHandler } from "@fibr/builder";
+import {
+  classNames,
+  eventHandler,
+  type Block,
+  type BlockType,
+} from "@fibr/builder";
 import type { RangeTuple } from "fuse.js";
-import type { Block } from "../../types";
 import { highlightMatches } from "../utils";
 
 export type PaletteCard = Block & {
