+++ packages/builder/shared/src/components/Palette/PaletteCardOverlay.tsx
@@ -1,5 +1,5 @@
 import { DragOverlay, useDndContext } from "@dnd-kit/core";
-import type { Block } from "../../types";
+import type { Block } from "@fibr/builder";
 import { PaletteCard } from "./PaletteCard";
 
 export function PaletteCardOverlay() {
