+++ packages/builder/shared/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./components";
 export * from "./providers";
 export * from "./resolvers";
-export * from "./types";
