--- packages/builder/shared/src/providers/blocks/index.ts
@@ -1,2 +1,0 @@
-export * from "./provider";
-export * from "./store";
