+++ packages/builder/shared/src/providers/index.ts
@@ -1,2 +1,1 @@
-export * from "./blocks";
 export * from "./shiki";
