--- packages/builder/shared/src/types/block.ts
@@ -1,10 +1,0 @@
-import type { IconType } from "react-icons/lib";
-
-export type Block<T extends Record<string, unknown> = Record<string, unknown>> =
-  {
-    type: string;
-    label: string;
-    icon: IconType;
-    private?: boolean;
-    presets?: T;
-  };
