--- packages/builder/shared/src/types/config.ts
@@ -1,10 +1,0 @@
-import type { ReactNode } from "react";
-
-export type Config = {
-  builder: () => ReactNode;
-  settings?: () => ReactNode;
-  actions?: () => ReactNode;
-  metadata?: {
-    node_type?: "edge" | "node";
-  };
-};
