--- packages/builder/shared/src/types/index.ts
@@ -1,2 +1,0 @@
-export type { Block } from "./block";
-export type { Config } from "./config";
